Add files via upload
Brain Tumor Detection in MRI images by tissue feature extraction and Machine Learning approaches
Introduction: Brain tumor detection is a challenging task that requires accurate diagnosis for effective treatment planning. Magnetic resonance imaging (MRI) has proven to be a valuable tool in detecting brain tumors. In this project, we propose to use machine learning and tissue feature extraction techniques to improve the accuracy of brain tumor detection in MRI images. We will use a dataset of MRI images from Kaggle that contains 98 healthy images and 155 images with tumors.
Objectives: The main objective of this project is to develop a machine learning model that can accurately detect brain tumors in MRI images. To achieve this objective, we will employ the following specific goals:
1.	To extract tissue features from MRI images using clustering techniques.
2.	To homogenize the images to ensure that they have the same dimensions.
3.	To use support vector machine (SVM) to classify the images as either healthy or tumor.
4.	To evaluate the performance of the proposed approach using various evaluation metrics.
Methodology: The proposed methodology involves the following steps:
1.	Data preprocessing: The first step will involve pre-processing the MRI images to homogenize the dimensions of the images. This will ensure that we can use clustering techniques to extract tissue features effectively.
2.	Tissue feature extraction: We will use clustering techniques to extract tissue features from the MRI images. Clustering techniques will be used to identify areas in the MRI images with the possibility of a tumor. These tissue features will then be used as input to SVM for classification.
3.	SVM classification: We will use SVM to classify the MRI images as either healthy or tumor. The extracted tissue features will be used as input to SVM for classification. The performance of the SVM classifier will be evaluated using various evaluation metrics.
4.	Performance evaluation: The performance of the proposed approach will be evaluated using various evaluation metrics such as accuracy, precision, recall, and F1-score. The evaluation will be carried out on a separate validation dataset to ensure that the results are reliable.
Expected Results: The proposed approach is expected to improve the accuracy of brain tumor detection in MRI images. By using tissue feature extraction and SVM classification, we expect to achieve higher accuracy compared to using the whole MRI image for classification. The evaluation metrics will be used to quantify the performance of the proposed approach.
Conclusion: In conclusion, this project proposes a new approach to improve the accuracy of brain tumor detection in MRI images. The proposed approach involves the use of tissue feature extraction and SVM classification. The performance of the proposed approach will be evaluated using various evaluation metrics. We expect the proposed approach to achieve higher accuracy compared to using the whole MRI image for classification.